+++ protected/humhub/modules/content/models/Content.php
@@ -247,11 +247,17 @@ class Content extends ActiveRecord implements Movable, ContentOwner, SoftDeletab
         }
 
         if ($insert || array_key_exists('state', $changedAttributes)) {
+            $previousState = $changedAttributes['state'] ?? null;
             $this->trigger(self::EVENT_STATE_CHANGED, new ContentStateEvent([
                 'content' => $this,
                 'newState' => $this->state,
-                'previousState' => $changedAttributes['state'] ?? null
+                'previousState' => $previousState
             ]));
+
+            $model = $this->getPolymorphicRelation();
+            if ($model instanceof ContentActiveRecord) {
+                $model->afterStateChange($this->state, $previousState);
+            }
         }
 
         if ($this->state === static::STATE_PUBLISHED) {