+++ CHANGELOG.md
@@ -4,7 +4,7 @@ HumHub Changelog
 1.14.1 (Unreleased)
 -------------------
 - Fix #6251: Emulate execution on `readable()` content 
-- Fix #6252: Fix event trigger for `Content::EVENT_STATE_CHANGED`
+- Enh #6252: Implement new method to handle changing of content active record state
 
 1.14.0 (April 20, 2023)
 -----------------------